# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "チャット駆動の予定登録を行う CalendarRabbit のバックエンドAPI（単一ユーザー・認証なし・VPN前提）。",
        "title": "CalendarRabbit API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/calendar/events": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "イベント一覧を取得する（from/to 指定で期間取得）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "期間開始(RFC3339)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "期間終了(RFC3339)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.eventResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "イベントを作成する",
                "parameters": [
                    {
                        "description": "イベント",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.eventRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.eventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/calendar/events/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "イベントを取得する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "イベントID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.eventResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "イベントを更新する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "イベントID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "イベント",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.eventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.eventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "calendar"
                ],
                "summary": "イベントを削除する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "イベントID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/chat/conversations": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "会話履歴を取得する（単一連続スレッド）",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.conversationResponse"
                        }
                    }
                }
            }
        },
        "/api/chat/messages": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "チャットメッセージを送信する",
                "parameters": [
                    {
                        "description": "メッセージ",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.sendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.sendMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/chat/proposals/{id}/approve": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "予定案を承認してカレンダーに登録する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "予定案ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "編集後内容(任意)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handler.approveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.eventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/chat/proposals/{id}/reject": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "予定案を却下する",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "予定案ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.proposalResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "ヘルスチェック",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/settings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "アプリ設定を取得する",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.settingsResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "アプリ設定を更新する",
                "parameters": [
                    {
                        "description": "設定",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.settingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.settingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.approveRequest": {
            "type": "object",
            "properties": {
                "edited": {
                    "$ref": "#/definitions/handler.eventRequest"
                }
            }
        },
        "handler.candidateResponse": {
            "type": "object",
            "properties": {
                "all_day": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "source_url": {
                    "type": "string"
                },
                "starts_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "handler.conversationResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.messageResponse"
                    }
                },
                "proposals": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.proposalResponse"
                    }
                }
            }
        },
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                }
            }
        },
        "handler.eventRequest": {
            "type": "object",
            "properties": {
                "all_day": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "source_url": {
                    "type": "string"
                },
                "starts_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "handler.eventResponse": {
            "type": "object",
            "properties": {
                "all_day": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "source_url": {
                    "type": "string"
                },
                "starts_at": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "handler.messageResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "handler.proposalResponse": {
            "type": "object",
            "properties": {
                "all_day": {
                    "type": "boolean"
                },
                "calendar_event_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "ends_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "source_url": {
                    "type": "string"
                },
                "starts_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "handler.sendMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "handler.sendMessageResponse": {
            "type": "object",
            "properties": {
                "assistant_message": {
                    "$ref": "#/definitions/handler.messageResponse"
                },
                "candidates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.candidateResponse"
                    }
                },
                "proposal": {
                    "$ref": "#/definitions/handler.proposalResponse"
                },
                "user_message": {
                    "$ref": "#/definitions/handler.messageResponse"
                }
            }
        },
        "handler.settingsRequest": {
            "type": "object",
            "properties": {
                "timezone": {
                    "type": "string"
                }
            }
        },
        "handler.settingsResponse": {
            "type": "object",
            "properties": {
                "timezone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}